refactor/move provider from testing folder
