+++ g4f/Provider/Providers/Wewordle.py
@@ -4,7 +4,7 @@ import json
 import random
 import time
 import string
-# from ...typing import sha256, Dict, get_type_hints
+from ...typing import sha256, Dict, get_type_hints
 
 url = "https://wewordle.org/gptapi/v1/android/turbo"
 model = ['gpt-3.5-turbo']
@@ -68,30 +68,6 @@ def _create_completion(model: str, messages: list, stream: bool, **kwargs):
         print(f"Error Occurred::{response.status_code}")
         return None
 
-# params = f'g4f.Providers.{os.path.basename(__file__)[:-3]} supports: ' + \
-#     '(%s)' % ', '.join(
-#         [f"{name}: {get_type_hints(_create_completion)[name].__name__}" for name in _create_completion.__code__.co_varnames[:_create_completion.__code__.co_argcount]])
-
-
-# Temporary For ChatCompletion Class
-class ChatCompletion:
-    @staticmethod
-    def create(model: str, messages: list, provider: None or str, stream: bool = False, auth: str = False, **kwargs):
-        kwargs['auth'] = auth
-
-        if provider and needs_auth and not auth:
-            print(
-                f'ValueError: {provider} requires authentication (use auth="cookie or token or jwt ..." param)', file=sys.stderr)
-            sys.exit(1)
-
-        try:
-            
-
-            return (_create_completion(model, messages, stream, **kwargs)
-                    if stream else ''.join(_create_completion(model, messages, stream, **kwargs)))
-        except TypeError as e:
-            print(e)
-            arg: str = str(e).split("'")[1]
-            print(
-                f"ValueError: {provider} does not support '{arg}' argument", file=sys.stderr)
-            sys.exit(1)
+params = f'g4f.Providers.{os.path.basename(__file__)[:-3]} supports: ' + \
+    '(%s)' % ', '.join(
+        [f"{name}: {get_type_hints(_create_completion)[name].__name__}" for name in _create_completion.__code__.co_varnames[:_create_completion.__code__.co_argcount]])
