--- testing/aiservice/README.md
@@ -1,2 +1,0 @@
-https://github.com/xtekky/gpt4free/issues/40#issuecomment-1629152431
-probably gpt-3.5
